로그인 시 수집한 위치정보로 세션에 업데이트
