+++ src/main/resources/application.properties
@@ -17,7 +17,7 @@ file.upload-dir-item=./src/main/resources/static/itemImages/
 
 #kakao.restApiKey=8d3faa5d5d173a056d112814d44db9c8
 google.api.key=AIzaSyAGgvPzN57_n5tGKb1sNvZJgg6Ogiv1CkY
-kakao.api.javascript.key=b6cde7a2302dca191e1132a0becc2f9d
+kakao.restApi.Key=eb8594daeeea0309b1b00616b4895fea
 
 
 spring.jpa.hibernate.ddl-auto=none
